chore(dashboard): remove accessibility testing and its CI job
Drops the Dashboard accessibility job, the 18 axe e2e harnesses, the two
axe npm scripts, and the @axe-core/playwright dependency.

The visual audit tools are kept: they are screenshot/measurement tools,
not accessibility ones, and only shared the harness's bundle-serving
half. That half moves to e2e/static-server.ts under a name that says what
it does, and stories/audit.ts and topography-audit.ts lose their axe
sections while keeping their visual work.
