+++ .github/workflows/ci.yml
@@ -576,63 +576,6 @@ jobs:
       - name: Run dashboard integration tests
         run: cargo nextest run --test dashboard_api_test --features test-transport --no-fail-fast --locked
 
-  # WCAG 2.2 AA is an acceptance criterion, so the accessibility scans run in
-  # CI rather than only on a developer's machine. Both harnesses exit nonzero
-  # on an axe violation, an uncaught page error, or a failed state assertion —
-  # a clean scan of a falsified reading is not a pass.
-  #
-  # Non-blocking by owner decision 2026-07-31: axe findings stay visible in
-  # the run summary but no longer gate the branch. Zero-axe remains a release
-  # acceptance criterion checked before release, not per push.
-  # NOTHING may declare `needs: dashboard-accessibility`.
-  dashboard-accessibility:
-    name: Dashboard accessibility
-    # `needs:` keeps this job from reporting a second red when the real fault
-    # is that the dashboard bundle never built or its unit tests failed.
-    needs: [dashboard, scope-gate]
-    if: ${{ needs.scope-gate.outputs.run-heavy == 'true' }}
-    runs-on: ubuntu-latest
-    continue-on-error: true
-    # Locally axe:audit takes ~3.3 min and axe:explorer ~1.3 min, and each one
-    # runs its OWN full `rsbuild build` (see dashboard/e2e/axe-harness.ts). At
-    # ~3x for a 2-core runner rendering 2 themes x 3 widths that is ~14 min,
-    # plus ~3 min for the Chromium and system-dependency install and ~1 min
-    # for npm ci. 30 leaves headroom for a slow runner while still cutting off
-    # a hung browser or wedged static server.
-    timeout-minutes: 30
-    steps:
-      - uses: actions/checkout@v7
-
-      - uses: actions/setup-node@v6
-        with:
-          node-version: "22"
-          cache: npm
-          cache-dependency-path: dashboard/package-lock.json
-
-      - name: Install dashboard dependencies
-        working-directory: dashboard
-        run: npm ci
-
-      - name: Cache Playwright browsers
-        uses: actions/cache@v6
-        with:
-          path: ~/.cache/ms-playwright
-          key: playwright-${{ runner.os }}-${{ hashFiles('dashboard/package-lock.json') }}
-          restore-keys: |
-            playwright-${{ runner.os }}-
-
-      - name: Install Playwright Chromium
-        working-directory: dashboard
-        run: npx playwright install --with-deps chromium
-
-      - name: Accessibility scan (workspaces)
-        working-directory: dashboard
-        run: npm run axe:audit
-
-      - name: Accessibility scan (Explorer)
-        working-directory: dashboard
-        run: npm run axe:explorer
-
   hermes-integration:
     name: Hermes integration (stock)
     needs: scope-gate
